(chore): 사건 조회 common 빌드 수정
